UpsertBusStop method added. Request may need to be changed to fit more specifically to each request, since GetPeopleCount may only need the Id

+++ backend/p8mobility.shared/BusStop.cs
@@ -5,8 +5,8 @@ namespace p8_shared;
 public class BusStop
 {
     public Guid Id { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
     public int PeopleCount { get; set; }
 
     public BusStop()
@@ -14,7 +14,7 @@ public class BusStop
         
     }
 
-    public BusStop(Guid id, decimal latitude, decimal longitude)
+    public BusStop(Guid id, double latitude, double longitude)
     {
         Id = id;
         Latitude = latitude;
